extract `pangolin-constants`
