+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
 	"bin/node/cli",
 	"bin/node/primitives",
 	"bin/node/runtime/pangolin",
+	"bin/node/runtime/pangolin/src/constants",
 	"client/cli",
 	"client/dvm/db",
 	"client/dvm/mapping-sync",
