+++ bin/node/runtime/pangolin/src/lib.rs
@@ -22,43 +22,6 @@
 // `construct_runtime!` does a lot of recursion and requires us to increase the limit to 256.
 #![recursion_limit = "256"]
 
-pub mod constants {
-	// --- substrate ---
-	use sp_staking::SessionIndex;
-	// --- darwinia ---
-	use crate::*;
-
-	pub const NANO: Balance = 1;
-	pub const MICRO: Balance = 1_000 * NANO;
-	pub const MILLI: Balance = 1_000 * MICRO;
-	pub const COIN: Balance = 1_000 * MILLI;
-
-	pub const CAP: Balance = 10_000_000_000 * COIN;
-	pub const TOTAL_POWER: Power = 1_000_000_000;
-
-	// Time is measured by number of blocks.
-	pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);
-	pub const HOURS: BlockNumber = 60 * MINUTES;
-	pub const DAYS: BlockNumber = 24 * HOURS;
-
-	pub const MILLISECS_PER_BLOCK: Moment = 6000;
-	// NOTE: Currently it is not possible to change the slot duration after the chain has started.
-	//       Attempting to do so will brick block production.
-	pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
-	// NOTE: Currently it is not possible to change the epoch duration after the chain has started.
-	//       Attempting to do so will brick block production.
-	pub const BLOCKS_PER_SESSION: BlockNumber = 10 * MINUTES;
-	pub const SESSIONS_PER_ERA: SessionIndex = 3;
-
-	// 1 in 4 blocks (on average, not counting collisions) will be primary babe blocks.
-	pub const PRIMARY_PROBABILITY: (u64, u64) = (1, 4);
-
-	pub const fn deposit(items: u32, bytes: u32) -> Balance {
-		items as Balance * 20 * COIN + (bytes as Balance) * 100 * MICRO
-	}
-}
-pub use constants::*;
-
 pub mod pallets;
 pub use pallets::*;
 
@@ -220,6 +183,7 @@ pub mod wasm {
 pub use wasm::*;
 
 pub use darwinia_staking::StakerStatus;
+pub use pangolin_constants::*;
 
 // --- crates ---
 use codec::{Decode, Encode};
