+++ bin/node/runtime/pangolin/src/pallets/multisig.rs
@@ -5,9 +5,9 @@ use crate::*;
 
 frame_support::parameter_types! {
 	// One storage item; key size is 32; value is size 4+4+16+32 bytes = 56 bytes.
-	pub const DepositBase: Balance = constants::deposit(1, 88);
+	pub const DepositBase: Balance = pangolin_constants::deposit(1, 88);
 	// Additional storage item size of 32 bytes.
-	pub const DepositFactor: Balance = constants::deposit(0, 32);
+	pub const DepositFactor: Balance = pangolin_constants::deposit(0, 32);
 	pub const MaxSignatories: u16 = 100;
 }
 impl Config for Runtime {
